# docs/swagger.yaml
definitions:
  controllers.InputLogin:
    properties:
      email:
        type: string
      password:
        type: string
    type: object
  controllers.InputUser:
    properties:
      address:
        type: string
      email:
        type: string
      name:
        type: string
      password:
        type: string
      phone:
        type: string
    type: object
...
  model.ErrorResponse:
    properties:
      message:
        type: string
      status:
        type: integer
    type: object
...
  model.User:
    properties:
      address:
        type: string
      email:
        type: string
      id:
        type: integer
      name:
        type: string
      password:
        type: string
      phone:
        type: string
      usertype:
        type: integer
    type: object
  model.UserResponse:
    properties:
      data:
        $ref: '#/definitions/model.User'
      message:
        type: string
      status:
        type: integer
    type: object
...
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
paths:
...
  /login:
    post:
      description: login for registered users.
      parameters:
      - description: User's login data
        in: body
        name: Body
        required: true
        schema:
          $ref: '#/definitions/controllers.InputLogin'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.ErrorResponse'
      summary: login user.
      tags:
      - Users
...
  /register:
    post:
      description: insert user and it use for register user.
      parameters:
      - description: User's data
        in: body
        name: Body
        required: true
        schema:
          $ref: '#/definitions/controllers.InputUser'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.UserResponse'
      summary: insert user.
      tags:
      - Users
...
  /user:
    put:
      description: change the data of the user who is currently logged in.
      parameters:
      - description: User's data
        in: body
        name: Body
        required: true
        schema:
          $ref: '#/definitions/controllers.InputUser'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.UserResponse'
      summary: update user.
      tags:
      - Users
  /user/{id}:
    delete:
      description: Delete user by id and admin only can use it.
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.ErrorResponse'
      summary: delete user.
      tags:
      - Users
...
swagger: "2.0"